requirements.txt
Flask
pdfplumber
python-docx
Pillow
pytesseract

